Moved GameSocketHandler to main directory
